Massive Changes
For the most part, the calculator is now "functional". I haven't found a recipe that enters an infinite loop yet, and most of them calculate reasonably well.
Oil is a pain, however. Next will be designing an interface that allows the user to change the priorities of coal and oil.