Proliferators
Hacked the modules system to add proliferator functionality. By replacing beacons with a slider to toggle between productivity and speed, the system mimics that of the game surprisingly well.
Given this was the last actual feature I had planned, this calculator is functionally done. All that's left is checking/fixing the power stats and updating the docs.