On-site proliferation assembling
Made it so that proliferator modules are treated similarly to fuel in the Factorio Calculator - that is, proliferator modules are added to the totals and incorporated into the ingredients list. This is messy, far from performant, and generally makes me sad. But it works.
Also added self-proliferation where proliferator items are sprayed to increase that item's sprays. The implementation is a bit hacky, but again, it works.
Next steps will include fine tuning the tinyNum for performance, and probably adding in a "Calculating" message to let the user know why the page has become unresponsive.
Also fixed the url fragmenting, so now we can save and share full recipes - buildings, proliferators, and all. The zips being longer than the url seems strange, and worth further investigation.